feat: rewrite kuby. (#26)
Closes #7.
Closes #5.

This essentially is a complete rewrite of
kuby in F# instead of typescript. F# provides
faster runtime and nicer language features and
it was a nice project to try the F# language.

Also, the direct need of `kubectl` is removed.
For the one part of the package that needs kubectl
(`deploy`) a bundled version of kubectl is used.

BREAKING CHANGE: The command `cleanup` is
removed. It was a highly destructive action
and therefore was removed.

BREAKING CHANGE: The command `kubectl` is
removed. The `kubectl` executable is included
in the compiled project and will be used from
there when necessary. To manage your local
`kubectl` versions, you can use
https://github.com/buehler/kubectl-version-manager

BREAKING CHANGE: The command `preview-deploy`
is removed. This was a command that was highly
opinionated and created more problems than it solved
It is removed in favor to CI-themed preview deployments.

BREAKING CHANGE: The commands `deploy`, `apply`,
and `prepare` are restructured. Instead of creating
three different commands, kuby now only knows `deploy`.
To create the same behaviour as in earlier versions,
deploy can be used with flags to output the prepared
yaml to a folder and use a dry-run flag to prevent
effective action on the cluster.

Signed-off-by: Christoph Bühler <christoph@smartive.ch>